Move post workspaces into workspace_handler
